backend/requirements.txt
flask
flask-cors
python-dotenv
flask-sqlalchemy
flask-migrate
fastapi
email-validator
passlib[bcrypt]
python-jose[cryptography]
python-multipart